fix or silence (msvc) compiler warnings about narrowing, unused formal parameters, and other minor stuff
(cherry picked from commit 5250a0d189c4b27946b9d746bca7c610c81ba12a)

+++ include/boost/python/suite/indexing/detail/indexing_suite_detail.hpp
@@ -602,7 +602,7 @@ namespace boost { namespace python { namespace detail {
             else {
                 long from = extract<long>( slice->start);
                 if (from < 0) // Negative slice index
-                    from += max_index;
+                    from += static_cast<long>(max_index);
                 if (from < 0) // Clip lower bounds to zero
                     from = 0;
                 from_ = boost::numeric_cast<Index>(from);
@@ -616,7 +616,7 @@ namespace boost { namespace python { namespace detail {
             else {
                 long to = extract<long>( slice->stop);
                 if (to < 0)
-                    to += max_index;
+                    to += static_cast<long>(max_index);
                 if (to < 0)
                     to = 0;
                 to_ = boost::numeric_cast<Index>(to);
